// src/api/model/formFieldFieldType.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Form API
 * API for creating and managing forms
 * OpenAPI spec version: 1.0.0
 */

/**
 * Type of the form field
 */
export type FormFieldFieldType =
  (typeof FormFieldFieldType)[keyof typeof FormFieldFieldType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FormFieldFieldType = {
  checkbox: "checkbox",
  radio: "radio",
  select: "select",
  textarea: "textarea",
  text: "text",
} as const;